Cache the resulting world generated from an MP3 file for performance.
Ensure to be quite defensive, and throw away files we don't like (including
ones from previous versions of the game if the format has changed, and corrupt
ones).

Speeds up performance substantially (for all but the first load).

Also added some logging here and a --verbose/-v flag to the desktop app to match.