Downsample from high bitrate songs, downsampling, but pre-calculate features first.
Allows for high quality obstacle generation, but smaller file size (and lower quality audio,
but that is the tradeoff with filesize). Rough estimate is that the resulting .apk for these
5 songs dropped from 35Mb -> 17Mb.

While it doesn't really matter that much for just these songs, it does mean we can add
many more songs into the future.