Refactor assets into an asset manager.
Still doesn't dispose of them properly, just loads once at the start of the game,
and also loads synchronously with no feedback to the user.

Despite all those issues, it is still better than before though, because before
we were constantly loading the texture atlas, constantly doing string lookups
in the texture atlas, and still loading synchronously and not disposing correctly.

So this is an improvement.