Enable uk and sv locales, and regenerate fonts.
This doesn't seem to actually export the skin from
SkinComposer. I guess becaues perhaps the skin itself
hasn't changed - just the dependent fonts - and it doesn't
know how to tell if that is the case or not. Either way,
will try to force the entire skin to re-export to the assets
folder in a subsequent commit.