Remove unused 'explain custom song' screen. Ensure legacy custom song is copied properly.
No need to explain where to put an MP3 file now, as we do this
by presenting a file chooser to the user and taking a copy of
the MP3 they select.

This also correctly copies the custom song from previous versions
to the new custom song location. Earlier attempts left it where
it was and pointed to it from the custom-world.json file. This
now copies it alongside other songs so it is not special, other
than having an ID of 'custom.mp3' so that achievements from the
past one are retained.