Initial attempt at 'delete custom levels' UX.
Players can long-press custom levels to prompt if they wish to delete.
There is a text hint at the bottom of the custom world selection screen
explaining they can long press (otherwise it is a bit obtuse).
After confirming, all achievements and high scores are deleted for that
level too, as well as the internal copy of the MP3 file and the JSON
level data.